pyproject.toml
[tool.poetry]
name = "market-spread-api"
version = "0.1.0"
description = ""
authors = ["Gabriel Anguita <gabriel.anguita.veas@gmail.com>"]
license = "GPL 3"
readme = "README.md"
packages = [{include = "market_spread_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
pydantic-settings = "^2.1.0"
aiohttp = {extras = ["speedups"], version = "^3.9.1"}
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
httpx = "^0.26.0"
aioresponses = "^0.7.6"
djangorestframework = "^3.14.0"
django = "^5.0"
python-dotenv = "^1.0.0"
djangorestframework-dataclasses = "^1.3.1"
adrf = "^0.1.2"
drf-spectacular = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
